# UniversityApp/swagger.yaml
openapi: 3.0.1
info:
  title: University API
  description: This is a university api documentation
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/group/{groupId}/addStudent/{studentId}:
    put:
      tags:
      - Group
      summary: "Adds student by = {studentId} to group by id = {groupId}"
      operationId: addStudent
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: studentId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/audience/{id}:
    get:
      tags:
      - Audience
      summary: Finds audience by id
      operationId: one
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Audience'
    delete:
      tags:
      - Audience
      summary: Deletes audience by id
      operationId: deleteOne_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/group/{id}:
    get:
      tags:
      - Group
      summary: Finds group by id
      operationId: one_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Group'
    delete:
      tags:
      - Group
      summary: Deletes group by id
      operationId: deleteOne_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /api/teacher/{id}:
    get:
      tags:
      - Teacher
      summary: Finds teacher by id
      operationId: one_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Teacher'
    delete:
      tags:
      - Teacher
      summary: Deletes teacher by id
      operationId: deleteOne
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/lecture/{id}:
    get:
      tags:
      - Lecture
      summary: Finds lecture by id
      operationId: one_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Lecture'
    delete:
      tags:
      - Lecture
      summary: Deletes lecture by id
      operationId: deleteOne_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /group/deleteFromFaculty:
    get:
      operationId: deleteFromFaculty
      parameters:
      - name: groupId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: facultyId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /api/lecture/search/{dateTime}:
    get:
      tags:
      - Lecture
      summary: Finds lectures by date
      operationId: searchByName
      parameters:
      - name: dateTime
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lecture'
  /api/faculty/{id}:
    get:
      tags:
      - Faculty
      summary: Finds faculty by id
      operationId: one_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Faculty'
    delete:
      tags:
      - Faculty
      summary: Deletes faculty by id
      operationId: deleteOne_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/course/search/{name}:
    get:
      tags:
      - Course
      summary: Finds courses by name
      operationId: searchByName_1
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
...
  /api/student/{id}:
    get:
      tags:
      - Student
      summary: Finds student by id
      operationId: one_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Student'
    delete:
      tags:
      - Student
      summary: Deletes student by id
      operationId: deleteOne_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /group/deleteFromLecture:
    get:
      operationId: deleteFromLecture
      parameters:
      - name: groupId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: lectureId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /api/group/{groupId}/deleteFromFaculty:
    delete:
      tags:
      - Group
      summary: Deletes group from its faculty
      operationId: deleteFromFaculty_1
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /teacher/deleteFromFaculty:
    get:
      operationId: deleteFromFaculty_2
      parameters:
      - name: teacherId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: facultyId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/group/{groupId}/fromLecture/{lectureId}:
    delete:
      tags:
      - Group
      summary: "Deletes group by id = {groupId} from lecture by id = {lectureId}"
      operationId: deleteFromLecture_1
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: lectureId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/course/{id}:
    get:
      tags:
      - Course
      summary: Finds course by id
      operationId: one_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Course'
    delete:
      tags:
      - Course
      summary: Deletes course by id
      operationId: deleteOne_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/teacher/{teacherId}/fromFaculty/{facultyId}:
    delete:
      tags:
      - Teacher
      summary: "Deletes teacher by id = {teacherId} from faculty by id = {facultyId}"
      operationId: deleteFromFaculty_3
      parameters:
      - name: teacherId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: facultyId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /api/group/search/{name}:
    get:
      tags:
      - Group
      summary: Finds groups by name
      operationId: searchByName_2
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
  /api/faculty/search/{name}:
    get:
      tags:
      - Faculty
      summary: Finds faculties by name
      operationId: searchByName_3
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Faculty'
...
  /group/addStudent:
    post:
      operationId: addStudent_1
      parameters:
      - name: groupId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CrmObject'
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    Audience:
      required:
      - audienceNumber
      - seats
      type: object
      properties:
        id:
          type: integer
          format: int32
        audienceNumber:
          maximum: 999
          minimum: 1
          type: integer
          format: int32
        seats:
          maximum: 200
          minimum: 10
          type: integer
          format: int32
        hasProjector:
          type: boolean
    Faculty:
      required:
      - description
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 2147483647
          minLength: 3
          type: string
        description:
          maxLength: 2147483647
          minLength: 10
          type: string
    Group:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 2147483647
          minLength: 2
          type: string
    CrmObject:
      required:
      - id
      type: object
      properties:
        id:
          minimum: 1
          type: integer
          format: int32
    Course:
      required:
      - description
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 2147483647
          minLength: 3
          type: string
        description:
          maxLength: 2147483647
          minLength: 10
          type: string
    Student:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          maxLength: 2147483647
          minLength: 2
          type: string
        lastName:
          maxLength: 2147483647
          minLength: 2
          type: string
        email:
          type: string
    Lecture:
      type: object
      properties:
        id:
          type: integer
          format: int32
        course:
          $ref: '#/components/schemas/Course'
        audience:
          $ref: '#/components/schemas/Audience'
        teacher:
          $ref: '#/components/schemas/Teacher'
        dateTime:
          type: string
          format: date-time
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    Teacher:
      required:
      - firstName
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          maxLength: 2147483647
          minLength: 2
          type: string
        lastName:
          maxLength: 2147483647
          minLength: 2
          type: string
...